requirements.txt
selenium
python-dotenv
openai
pyodbc
azure-identity